azure-functions/requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.12.13; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
async-timeout==5.0.1; python_version >= '3.8'
attrs==25.3.0; python_version >= '3.8'
azure-core==1.34.0; python_version >= '3.9'
azure-data-tables==12.7.0; python_version >= '3.9'
azure-functions==1.23.0; python_version >= '3.7'
azure-storage-blob==12.25.1; python_version >= '3.8'
azure-storage-queue==12.12.0; python_version >= '3.8'
beautifulsoup4==4.13.4; python_full_version >= '3.7.0'
certifi==2025.6.15; python_version >= '3.7'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.2; python_version >= '3.7'
cryptography==45.0.4; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
frozenlist==1.7.0; python_version >= '3.9'
idna==3.10; python_version >= '3.6'
isodate==0.7.2; python_version >= '3.7'
lxml==5.4.0; python_version >= '3.6'
markupsafe==3.0.2; python_version >= '3.9'
multidict==6.5.0; python_version >= '3.9'
propcache==0.3.2; python_version >= '3.9'
pycparser==2.22; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2025.2
requests==2.32.4; python_version >= '3.8'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
soupsieve==2.7; python_version >= '3.8'
typing-extensions==4.14.0; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
yarl==1.20.1; python_version >= '3.9'
